requirements.txt
tensorflow
streamlit
numpy
matplotlib
Pillow
